requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
pywinpty>=2.0
